pyproject.toml
[project]
name = "talkops"
version = "1.0.6"
authors = [{ name = "PicoUX" }]
description = "TalkOps SDK"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "jinja2>=3.1.6",
    "requests>=2.32.3",
    "sseclient-py>=1.8.0"
]

[project.urls]
Homepage = "https://talkops.app"
Issues = "https://github.com/talkops/sdk-python/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"talkops" = ["data/*"]

[tool.setuptools.packages.find]
where = ["."]
